+++ src/geometry/linear_extrude.cc
@@ -720,13 +720,13 @@ std::unique_ptr<Geometry> extrudePolygon(const LinearExtrudeNode& node, const Po
         }
       } else {
         builder.beginPolygon(topvertices.size());  // bottom
-        for (int i = topvertices.size() - 1; i >= 0; i--) builder.addVertex(topvertices[i]);
+        for (size_t i = topvertices.size() - 1; i >= 0; i--) builder.addVertex(topvertices[i]);
         builder.endPolygon();
       }
       botvertices = topvertices;
     }
     builder.beginPolygon(botvertices.size());  // top
-    for (int i = 0; i < botvertices.size(); i++) builder.addVertex(botvertices[i]);
+    for (size_t i = 0; i < botvertices.size(); i++) builder.addVertex(botvertices[i]);
     builder.endPolygon();
 
     return builder.build();
